// elearning-frontend/src/client-sdk/models/CommentResource.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { LessonNotification } from './LessonNotification';
export type CommentResource = {
  lesson: LessonNotification;
  id: string;
  content: string;
};