// elearning-frontend/src/client-sdk/models/NotificationType.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export enum NotificationType {
  NEW_REVIEW = 'NEW_REVIEW',
  NEW_COMMENT = 'NEW_COMMENT',
  NEW_REPLY = 'NEW_REPLY',
  NEW_ENROLLMENT = 'NEW_ENROLLMENT',
}