// elearning-frontend/src/client-sdk/models/ReplyResource.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { LessonNotification } from './LessonNotification';
export type ReplyResource = {
  lesson: LessonNotification;
  id: string;
  content: string;
  parentId: string | null;
};