// elearning-frontend/src/client-sdk/models/ReviewResource.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { CourseNotification } from './CourseNotification';
import type { Rating } from './Rating';
export type ReviewResource = {
  rating: Rating;
  course: CourseNotification;
  id: string;
  content: string | null;
};